requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
ujson==5.5.0
urllib3==1.26.12